refactor: 큇메뉴, 검색 드로어 kks레이아웃에 추가 / 드로어 display 조건 변경
