+++ apps/web-kkshow/pages/index.tsx
@@ -1,19 +1,17 @@
 import { Box } from '@chakra-ui/react';
 import { kkshowFooterLinkList } from '@project-lc/components-constants/footerLinks';
 import { CommonFooter } from '@project-lc/components-layout/CommonFooter';
-import { KkshowNavbar } from '@project-lc/components-web-kkshow/KkshowNavbar';
 import { KkshowMainBestBroadcaster } from '@project-lc/components-web-kkshow/main/KkshowMainBestBroadcaster';
 import { KkshowMainBestLive } from '@project-lc/components-web-kkshow/main/KkshowMainBestLive';
 import { KkshowMainCarousel } from '@project-lc/components-web-kkshow/main/KkshowMainCarousel';
 import { KKshowMainExternLinks } from '@project-lc/components-web-kkshow/main/KKshowMainExternLinks';
 import { KkshowLiveTeaser } from '@project-lc/components-web-kkshow/main/KkshowMainLiveTeaser';
 import { KkshowMainPlusFriend } from '@project-lc/components-web-kkshow/main/KkshowMainPlusFriend';
-import { BottomQuickMenu } from '@project-lc/components-shared/BottomQuickMenu';
 import { getKkshowMain, kkshowMainQueryKey } from '@project-lc/hooks';
 import { createQueryClient } from '@project-lc/utils-frontend';
 import { GetStaticProps } from 'next';
 import { dehydrate, DehydratedState } from 'react-query';
-import { MobileSearchDrawer } from '@project-lc/components-web-kkshow/search/MobileSearchDrawer';
+import { KkshowLayout } from '@project-lc/components-web-kkshow/KkshowLayout';
 
 interface KkshowIndexProps {
   dehydratedState: DehydratedState;
@@ -65,17 +63,16 @@ export default function Index(): JSX.Element {
         zIndex={1}
       />
 
-      <KkshowNavbar />
-      <KkshowMainCarousel />
-      <KkshowLiveTeaser />
-      <KkshowMainBestLive />
-      <KkshowMainPlusFriend />
-      <KkshowMainBestBroadcaster />
+      <KkshowLayout>
+        <KkshowMainCarousel />
+        <KkshowLiveTeaser />
+        <KkshowMainBestLive />
+        <KkshowMainPlusFriend />
+        <KkshowMainBestBroadcaster />
 
-      <KKshowMainExternLinks mb={-4} bgColor="blue.900" color="whiteAlpha.900" />
-      <CommonFooter footerLinkList={kkshowFooterLinkList} />
-      <BottomQuickMenu />
-      <MobileSearchDrawer />
+        <KKshowMainExternLinks mb={-4} bgColor="blue.900" color="whiteAlpha.900" />
+        <CommonFooter footerLinkList={kkshowFooterLinkList} />
+      </KkshowLayout>
     </Box>
   );
 }
