+++ libs/components-web-kkshow/src/lib/KkshowLayout.tsx
@@ -1,6 +1,7 @@
 import { Box } from '@chakra-ui/react';
 import { BottomQuickMenu } from '@project-lc/components-shared/BottomQuickMenu';
 import { KkshowNavbar } from './KkshowNavbar';
+import { MobileSearchDrawer } from './search/MobileSearchDrawer';
 
 interface KkshowLayoutProps {
   children: React.ReactNode;
@@ -12,6 +13,7 @@ export function KkshowLayout({ children }: KkshowLayoutProps): JSX.Element {
       <KkshowNavbar />
       {children}
       <BottomQuickMenu />
+      <MobileSearchDrawer />
     </Box>
   );
 }
